feat(settings): full-screen settings page with instant back
Replace the settings modal (SettingsDialog) with a routed full-screen
page (/settings/:section) plus a left-nav layout. MainView is extracted
from App so App becomes the router shell.

Keep MainView mounted (hidden) while settings is open instead of
unmounting it, so returning to the workspace is instant — remounting
rebuilt the CodeMirror editor and H.264 decoder from scratch, which made
the back transition slow. Suppress MainView's global shortcuts while
settings is open so they don't fire from behind the overlay.
