+++ src/main.tsx
@@ -3,6 +3,7 @@
 
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { HashRouter } from "react-router-dom";
 import App from "./App";
 import "./styles/globals.css";
 
@@ -46,7 +47,9 @@ function hideSplash() {
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <App />
+    <HashRouter>
+      <App />
+    </HashRouter>
   </React.StrictMode>,
 );
